+++ tools/please_wit/generate/golang.go
@@ -158,6 +158,25 @@ func (g *GoGenerator) processTypeAliases(sb *strings.Builder, typeDefs []ast.Typ
 	}
 }
 
+func (g *GoGenerator) processResources(sb *strings.Builder, resources []ast.Resource) {
+	for _, res := range resources {
+		sb.WriteString(fmt.Sprintf("type %s interface {\n", ToPascalCase(res.Name)))
+		for _, fn := range res.Methods {
+			var params []string
+			for _, p := range fn.Params {
+				params = append(params, fmt.Sprintf("%s %s", ToCamelCase(p.Name), g.MapWitType(p.Type)))
+			}
+			retType := g.MapWitType(fn.Results)
+			retClause := ""
+			if retType != "" {
+				retClause = " " + retType
+			}
+			sb.WriteString(fmt.Sprintf("\t%s(%s)%s\n", ToPascalCase(fn.Name), strings.Join(params, ", "), retClause))
+		}
+		sb.WriteString("}\n\n")
+	}
+}
+
 func (g *GoGenerator) processInterface(sb *strings.Builder, iface ast.Interface) {
 	sb.WriteString(fmt.Sprintf("type %s interface {\n", ToPascalCase(iface.Name)))
 	for _, fn := range iface.Functions {
@@ -193,8 +212,21 @@ func (g *GoGenerator) generateGoCode(pkg *ast.Package, targetPackage string) str
 		// 3. Type aliases
 		g.processTypeAliases(&sb, iface.TypeDefs)
 
-		// 4. Interface declaration
-		g.processInterface(&sb, iface)
+		// 4. Resources
+		g.processResources(&sb, iface.Resources)
+
+		// 5. Interface declaration
+		hasResourceWithSameName := false
+		for _, res := range iface.Resources {
+			if ToPascalCase(res.Name) == ToPascalCase(iface.Name) {
+				hasResourceWithSameName = true
+				break
+			}
+		}
+
+		if !hasResourceWithSameName || len(iface.Functions) > 0 {
+			g.processInterface(&sb, iface)
+		}
 	}
 
 	return sb.String()